requirements.txt
fastapi
uvicorn[standard]
python-multipart
aiofiles
jinja2
python-dotenv
google-cloud-aiplatform
google-cloud-vision
bcrypt==4.0.1
httpx>=0.27.0
